upgradable stdlibs
Fixes #53983

Unfortunately, TOML has a design problem with the way it is used by Pkg,
so this does not yet work correctly. Need to finish changing uses of
loaded_packages to explicit_loaded_modules in loading to try to make it
less terribly broken.

Also currently one test is broken, as we need some more code
rearrangement to make it legal again for an `__init__` method to import code.