WIP: Refactor `Binding` data structures in preparation for partition
This is a re-worked extraction of #54654, adjusted to support the
new semantics arrived at over the course of that thread. Note that
this is the data-structure change only. The semantics in this PR
still match current master to the greatest extent possible. In
particular, this does not fully implement thread-safety of the
old semantics, since that works differently after #54654 (this
PR essentially implements the old semantics by mutating in place
data structures that are supposed to be immutable), though note
that there were some thread safety concerns in the existing also.

The core idea here is to split `Binding` in two: A new `Binding`
with minimal data and a `BindingPartition` that holds all data
that is world-age partitioned. In the present PR, these are always
in 1:1 correspondednce, but after #54654, there will be multiple
`BindingPartition`s for every `Binding`.

Essentially the `owner` and `ty` fields have been merged into
a new `restriction` field of `BindingPartition`, which may also
hold the value of a constant (consistent with the final semantics
reached in #54654). The non-partitioned binding->value field is
now used exclusively for non-constant globals. The disambiguation
for how to interpret the `restriction` field happens via flags.
`->imported` grew to 2 bits and can now be one of `NONE`/`IMPLICIT`/
`EXPLICIT`/`GUARD`. `GUARD` corresponds to the old `b->owner == NULL`
case. `NONE` corresponds to the old `b->owner == b` case, while
IMPLICIT/EXPLICIT correspond to `b->owner != b` and the old `imported`
flag. Other than that, the behavior of the flags is unchanged.

Additionally, fields are provided for `min_world`/`max_world`/`next`
and a `declared` flag in `Binding`, but these are unused in this PR
and prepratory only.

Because of above mentioned thread-safety concerns, I am not anticipating
merging this PR by itself, but I think it is a good review partition
to cut down on the number of changes in the main PRs.